Add Custom Modal Element For Use With Instructions
This commit adds functionality of a custom HTML element which renders a
modal to the screen. This custom element is based on #backdrop
Maximillian Schwarzmuller's Javascript tutorials which render the modal
via Javascript, CSS and using the shadow DOM.

This approach allows the rendering of a modal without importing
additional libraries such as Bootstrap, allowing the application to stay
lean.