Reorganise Structure Into Single Game Object
This commit reorganises the code to leverage object oriented
programming. From this commit forward the game will be organised into a
single object which has various methods to trigger each state of the
game (story, battle, instructions etc).

It is still not clear how the order of the three battles with story
links in between will be achieved but this will be decided later in the
process.