Add Mockups Of Full Score And Playback Logic
This commit adds a mockup for all four audio cues for the game. It additionally adds logic to loop each cue, as well as logic to
progress to the next cue when the appropriate trigger is reached. This is bundled into an audio object with a play method as well
as methods for each track which call upon the play method.