Add Update UI Module With Simple Function to Render UI
This commit adds a simple function which clears all UI elements and instead only renders the one UI element passed into the function. All UI elements are now
hosted in the index.html page, with a class of hidden, which removes them from the DOM. Upon calling the function, the class of hidden will be removed thereby
rendering the element to the screen and hiding all others.