Begin Battle Logic Addition
This commit adds the initial battle logic to allow battles with enemys
one and two to progress with just attack methods, building the initial
loop. This will for now be quite verbose and will be simplified later.
Additional commands will also be added and more complex logic will also
be added once the loop is functioning correctly