Implement Choices Array For Each Item In Story
This commit makes sure each stage in the story.js file has a choices
attribute which populates the list of actions that can be taken on each
screen. Each choice has an action which is the function that will be
called, and a next property which is the screen it will lead to.